+++ tests/utils/stdio.rs
